move debugging macros into haywire.h
