+++ src/webserver.c
@@ -4,10 +4,10 @@
 #pragma comment(lib, "Iphlpapi.lib")
 
 #include <assert.h>
-#include <stdarg.h>
 #include <stdio.h>
 #include <stdlib.h>
 
+#include "haywire.h"
 #include "uv.h"
 #include "http_parser.h"
 
@@ -18,25 +18,6 @@
     UVERR(err, msg); \
     exit(1); \
   }
-
-#ifndef DEBUG
-# ifdef NDEBUG
-#  define DEBUG 0
-# else
-#  define DEBUG 1
-# endif
-#endif
-
-void dbg_print(const char *fmt, ...)
-{
-    va_list args;
-    va_start(args, fmt);
-    vfprintf(stderr, fmt, args);
-    va_end(args);
-}
-
-#define DPUTS(msg) do { if (DEBUG) fputs(msg, stderr); } while (0)
-#define DPRINT(x) do { if (DEBUG) dbg_print x; } while (0)
 
 //"Connection: Keep-Alive\r\n" \
